Renamed component to email verification screen for clarity and changed back sign up screen text.
