+++ client/app/navigation/AuthNavigator.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 import LoginScreen from "../screens/auth/LoginScreen";
 import SignUpScreen from "../screens/auth/SignUpScreen";
 import WelcomeScreen from "../screens/auth/WelcomeScreen";
-import DigitVerificationScreen from "../screens/auth/DigitVerificationScreen";
+import EmailVerificationScreen from "../screens/auth/EmailVerificationScreen";
 
 const Stack = createStackNavigator();
 
@@ -17,7 +17,10 @@ const AuthNavigator = () => {
       <Stack.Screen name="Welcome" component={WelcomeScreen} />
       <Stack.Screen name="Log In" component={LoginScreen} />
       <Stack.Screen name="Sign Up" component={SignUpScreen} />
-      <Stack.Screen name="Digit Verification" component={DigitVerificationScreen} />
+      <Stack.Screen
+        name="Email Verification"
+        component={EmailVerificationScreen}
+      />
     </Stack.Navigator>
   );
 };
